# frontend/src/api/generated/openapi.json
{
  "components": {
    "schemas": {
...
      "AnimationClip": {
        "additionalProperties": false,
        "properties": {
          "duration": {
            "exclusiveMinimum": 0.0,
            "title": "Duration",
            "type": "number"
          },
          "engine_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Engine Version"
          },
          "events": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/ClipEvent"
            },
            "title": "Events",
            "type": "array"
          },
          "id": {
            "pattern": "^clip_[a-z0-9][a-z0-9_-]*$",
            "title": "Id",
            "type": "string"
          },
          "loop": {
            "default": false,
            "title": "Loop",
            "type": "boolean"
          },
          "loop_range": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Point2"
              },
              {
                "type": "null"
              }
            ]
          },
          "markers": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/ClipMarker"
            },
            "title": "Markers",
            "type": "array"
          },
          "name": {
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "scene_id": {
            "pattern": "^scene_[a-z0-9][a-z0-9_-]*$",
            "title": "Scene Id",
            "type": "string"
          },
          "source_plan_id": {
            "anyOf": [
              {
                "pattern": "^plan_[a-z0-9][a-z0-9_-]*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Plan Id"
          },
          "tracks": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/Track"
            },
            "title": "Tracks",
            "type": "array"
          }
        },
        "required": [
          "id",
          "scene_id",
          "name",
          "duration"
        ],
        "title": "AnimationClip",
        "type": "object"
      },
...
      "AssetManifestEntry": {
        "additionalProperties": false,
        "properties": {
          "display_name": {
            "default": "",
            "title": "Display Name",
            "type": "string"
          },
          "id": {
            "pattern": "^asset_[a-z0-9][a-z0-9_-]*$",
            "title": "Id",
            "type": "string"
          },
          "media_type": {
            "minLength": 1,
            "title": "Media Type",
            "type": "string"
          },
          "sha256": {
            "pattern": "^[0-9a-f]{64}$",
            "title": "Sha256",
            "type": "string"
          }
        },
        "required": [
          "id",
          "sha256",
          "media_type"
        ],
        "title": "AssetManifestEntry",
        "type": "object"
      },
...
      "CharacterDefinition": {
        "additionalProperties": false,
        "properties": {
          "attachments": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/AttachmentDefinition"
            },
            "title": "Attachments",
            "type": "array"
          },
          "id": {
            "pattern": "^char_[a-z0-9][a-z0-9_-]*$",
            "title": "Id",
            "type": "string"
          },
          "name": {
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "rig": {
            "$ref": "#/components/schemas/RigDefinition"
          }
        },
        "required": [
          "id",
          "name",
          "rig"
        ],
        "title": "CharacterDefinition",
        "type": "object"
      },
...
      "GenerationRecord": {
        "additionalProperties": false,
        "properties": {
          "attempts": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/GenerationAttempt"
            },
            "title": "Attempts",
            "type": "array"
          },
          "blueprint": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CharacterBlueprint"
              },
              {
                "type": "null"
              }
            ]
          },
          "builder_diagnostics": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/BuilderDiagnostic"
            },
            "title": "Builder Diagnostics",
            "type": "array"
          },
          "character_id": {
            "anyOf": [
              {
                "pattern": "^char_[a-z0-9][a-z0-9_-]*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Character Id"
          },
          "created_at": {
            "pattern": "^\\d{4}-\\d{2}-\\d{2}T\\d{2}:\\d{2}:\\d{2}(\\.\\d+)?(Z|[+-]\\d{2}:\\d{2})?$",
            "title": "Created At",
            "type": "string"
          },
          "failure_kind": {
            "$ref": "#/components/schemas/GenerationFailureKind",
            "default": "none"
          },
          "id": {
            "pattern": "^gen_[a-z0-9][a-z0-9_-]*$",
            "title": "Id",
            "type": "string"
          },
          "kind": {
            "$ref": "#/components/schemas/GenerationKind",
            "default": "character_blueprint"
          },
          "model_name": {
            "minLength": 1,
            "title": "Model Name",
            "type": "string"
          },
          "options": {
            "$ref": "#/components/schemas/GenerationOptionsRecord"
          },
          "outcome_detail": {
            "default": "",
            "title": "Outcome Detail",
            "type": "string"
          },
          "plan_id": {
            "anyOf": [
              {
                "pattern": "^plan_[a-z0-9][a-z0-9_-]*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Plan Id"
          },
          "plan_patch": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/MotionPlanPatch"
              },
              {
                "type": "null"
              }
            ]
          },
          "prompt_ids": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Prompt Ids",
            "type": "array"
          },
          "retryable": {
            "default": false,
            "title": "Retryable",
            "type": "boolean"
          },
          "status": {
            "$ref": "#/components/schemas/GenerationStatus"
          },
          "timing": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/GenerationTiming"
              },
              {
                "type": "null"
              }
            ]
          },
          "warnings": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Warnings",
            "type": "array"
          }
        },
        "required": [
          "id",
          "created_at",
          "model_name",
          "options",
          "status"
        ],
        "title": "GenerationRecord",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "MigrationReport": {
        "properties": {
          "applied": {
            "items": {
              "type": "string"
            },
            "title": "Applied",
            "type": "array"
          },
          "from_version": {
            "title": "From Version",
            "type": "string"
          },
          "to_version": {
            "title": "To Version",
            "type": "string"
          }
        },
        "required": [
          "from_version",
          "to_version"
        ],
        "title": "MigrationReport",
        "type": "object"
      },
...
      "ProjectCreate": {
        "properties": {
          "document": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ProjectDocument"
              },
              {
                "type": "null"
              }
            ]
          },
          "name": {
            "default": "Untitled Project",
            "minLength": 1,
            "title": "Name",
            "type": "string"
          }
        },
        "title": "ProjectCreate",
        "type": "object"
      },
      "ProjectDocument": {
        "additionalProperties": false,
        "properties": {
          "asset_manifest": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/AssetManifestEntry"
            },
            "title": "Asset Manifest",
            "type": "array"
          },
          "characters": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/CharacterDefinition"
            },
            "title": "Characters",
            "type": "array"
          },
          "clips": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/AnimationClip"
            },
            "title": "Clips",
            "type": "array"
          },
          "engine_version": {
            "pattern": "^\\d+\\.\\d+\\.\\d+$",
            "title": "Engine Version",
            "type": "string"
          },
          "format": {
            "pattern": "^rigstory-project$",
            "title": "Format",
            "type": "string"
          },
          "generation_records": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/GenerationRecord"
            },
            "title": "Generation Records",
            "type": "array"
          },
          "motion_plans": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/MotionPlan"
            },
            "title": "Motion Plans",
            "type": "array"
          },
          "project": {
            "$ref": "#/components/schemas/ProjectInfo"
          },
          "scenes": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/SceneDefinition"
            },
            "title": "Scenes",
            "type": "array"
          },
          "schema_version": {
            "pattern": "^\\d+\\.\\d+\\.\\d+$",
            "title": "Schema Version",
            "type": "string"
          }
        },
        "required": [
          "format",
          "schema_version",
          "engine_version",
          "project"
        ],
        "title": "ProjectDocument",
        "type": "object"
      },
      "ProjectImportRead": {
        "properties": {
          "document": {
            "$ref": "#/components/schemas/ProjectDocument"
          },
          "id_reassigned": {
            "title": "Id Reassigned",
            "type": "boolean"
          },
          "imported_assets": {
            "items": {
              "type": "string"
            },
            "title": "Imported Assets",
            "type": "array"
          },
          "migration": {
            "$ref": "#/components/schemas/MigrationReport"
          },
          "original_project_id": {
            "title": "Original Project Id",
            "type": "string"
          },
          "revision": {
            "title": "Revision",
            "type": "string"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "title": "Warnings",
            "type": "array"
          }
        },
        "required": [
          "document",
          "revision",
          "original_project_id",
          "id_reassigned",
          "migration"
        ],
        "title": "ProjectImportRead",
        "type": "object"
      },
...
      "ProjectRead": {
        "properties": {
          "document": {
            "$ref": "#/components/schemas/ProjectDocument"
          },
          "revision": {
            "title": "Revision",
            "type": "string"
          }
        },
        "required": [
          "document",
          "revision"
        ],
        "title": "ProjectRead",
        "type": "object"
      },
      "ProjectRevisionSummary": {
        "properties": {
          "id": {
            "title": "Id",
            "type": "string"
          },
          "sequence": {
            "title": "Sequence",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "sequence"
        ],
        "title": "ProjectRevisionSummary",
        "type": "object"
      },
      "ProjectSummary": {
        "properties": {
          "id": {
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "revision": {
            "title": "Revision",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "revision"
        ],
        "title": "ProjectSummary",
        "type": "object"
      },
      "ProjectUpdate": {
        "properties": {
          "document": {
            "$ref": "#/components/schemas/ProjectDocument"
          },
          "expected_revision": {
            "title": "Expected Revision",
            "type": "string"
          }
        },
        "required": [
          "document",
          "expected_revision"
        ],
        "title": "ProjectUpdate",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "RigStory Studio",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/projects": {
      "get": {
        "operationId": "list_projects_api_v1_projects_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ProjectSummary"
                  },
                  "title": "Response List Projects Api V1 Projects Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Projects",
        "tags": [
          "projects"
        ]
      },
      "post": {
        "operationId": "create_project_api_v1_projects_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjectCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Project",
        "tags": [
          "projects"
        ]
      }
    },
    "/api/v1/projects/import": {
      "post": {
        "description": "Import a portable archive uploaded as a raw ``application/zip`` body.",
        "operationId": "import_project_archive_api_v1_projects_import_post",
        "parameters": [
          {
            "in": "query",
            "name": "on_conflict",
            "required": false,
            "schema": {
              "default": "new_id",
              "enum": [
                "new_id",
                "fail"
              ],
              "title": "On Conflict",
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectImportRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Import Project Archive",
        "tags": [
          "projects"
        ]
      }
    },
    "/api/v1/projects/{project_id}": {
      "delete": {
        "operationId": "delete_project_api_v1_projects__project_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Project",
        "tags": [
          "projects"
        ]
      },
      "get": {
        "operationId": "get_project_api_v1_projects__project_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Project",
        "tags": [
          "projects"
        ]
      },
      "patch": {
        "operationId": "update_project_api_v1_projects__project_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjectUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Project",
        "tags": [
          "projects"
        ]
      }
    },
...
    "/api/v1/projects/{project_id}/duplicate": {
      "post": {
        "operationId": "duplicate_project_api_v1_projects__project_id__duplicate_post",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Duplicate Project",
        "tags": [
          "projects"
        ]
      }
    },
    "/api/v1/projects/{project_id}/export": {
      "get": {
        "description": "Download the project as a portable, checksummed archive.",
        "operationId": "export_project_archive_api_v1_projects__project_id__export_get",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/zip": {}
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Export Project Archive",
        "tags": [
          "projects"
        ]
      }
    },
    "/api/v1/projects/{project_id}/restore/{revision_id}": {
      "post": {
        "operationId": "restore_project_revision_api_v1_projects__project_id__restore__revision_id__post",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "revision_id",
            "required": true,
            "schema": {
              "title": "Revision Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Restore Project Revision",
        "tags": [
          "projects"
        ]
      }
    },
    "/api/v1/projects/{project_id}/revisions": {
      "get": {
        "operationId": "list_project_revisions_api_v1_projects__project_id__revisions_get",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ProjectRevisionSummary"
                  },
                  "title": "Response List Project Revisions Api V1 Projects  Project Id  Revisions Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Project Revisions",
        "tags": [
          "projects"
        ]
      }
    },
...
  }
}